Change type of properties in `CharacterSheet`

+++ nekoyume/Assets/_Scripts/TableData/Character/CharacterSheet.cs
@@ -16,15 +16,15 @@ namespace Nekoyume.TableData
             public int Id { get; private set; }
             public SizeType SizeType { get; private set; }
             public ElementalType ElementalType { get; private set; }
-            public int HP { get; private set; }
-            public int ATK { get; private set; }
-            public int DEF { get; private set; }
+            public decimal HP { get; private set; }
+            public decimal ATK { get; private set; }
+            public decimal DEF { get; private set; }
             public decimal CRI { get; private set; }
             public decimal HIT { get; private set; }
             public decimal SPD { get; private set; }
-            public int LvHP { get; private set; }
-            public int LvATK { get; private set; }
-            public int LvDEF { get; private set; }
+            public decimal LvHP { get; private set; }
+            public decimal LvATK { get; private set; }
+            public decimal LvDEF { get; private set; }
             public decimal LvCRI { get; private set; }
             public decimal LvHIT { get; private set; }
             public decimal LvSPD { get; private set; }
@@ -38,20 +38,20 @@ namespace Nekoyume.TableData
                 ElementalType = Enum.TryParse<ElementalType>(fields[2], out var elementalType)
                     ? elementalType
                     : ElementalType.Normal;
-                HP = int.TryParse(fields[3], out var hp) ? hp : 0;
-                ATK = int.TryParse(fields[4], out var damage) ? damage : 0;
-                DEF = int.TryParse(fields[5], out var defense) ? defense : 0;
+                HP = decimal.TryParse(fields[3], out var hp) ? hp : 0m;
+                ATK = decimal.TryParse(fields[4], out var damage) ? damage : 0m;
+                DEF = decimal.TryParse(fields[5], out var defense) ? defense : 0m;
                 CRI = decimal.TryParse(fields[6], out var cri) ? cri : 0m;
                 HIT = decimal.TryParse(fields[7], out var hit) ? hit : 0m;
                 SPD = decimal.TryParse(fields[8], out var spd) ? spd : 0m;
-                LvHP = int.TryParse(fields[9], out var lvHP) ? lvHP : 0;
-                LvATK = int.TryParse(fields[10], out var lvDamage) ? lvDamage : 0;
-                LvDEF = int.TryParse(fields[11], out var lvDefense) ? lvDefense : 0;
+                LvHP = decimal.TryParse(fields[9], out var lvHP) ? lvHP : 0m;
+                LvATK = decimal.TryParse(fields[10], out var lvDamage) ? lvDamage : 0m;
+                LvDEF = decimal.TryParse(fields[11], out var lvDefense) ? lvDefense : 0m;
                 LvCRI = decimal.TryParse(fields[12], out var lvCri) ? lvCri : 0m;
                 LvHIT = decimal.TryParse(fields[13], out var lvHit) ? lvHit : 0m;
                 LvSPD = decimal.TryParse(fields[14], out var lvSpd) ? lvSpd : 0m;
                 AttackRange = float.TryParse(fields[15], out var attackRange) ? attackRange : 1f;
-                RunSpeed = int.TryParse(fields[16], out var runSpeed) ? runSpeed : 1f;
+                RunSpeed = float.TryParse(fields[16], out var runSpeed) ? runSpeed : 1f;
             }
         }
